backend/requirements.txt
django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.6
google-adk>=1.25
python-dotenv>=1.0
